Move components provider into separate packege, add mapper test provider
