+++ src/test/java/easytests/core/mappers/providers/MapperComponentProvider.java
@@ -1,4 +1,4 @@
-package easytests.core.mappers.testschecker;
+package easytests.core.mappers.providers;
 
 import org.apache.ibatis.annotations.Mapper;
 import org.springframework.beans.factory.annotation.AnnotatedBeanDefinition;
@@ -8,8 +8,8 @@ import org.springframework.core.type.filter.AnnotationTypeFilter;
 /**
  * @author vkpankov
  */
-class InterfaceComponentProvider extends ClassPathScanningCandidateComponentProvider {
-    InterfaceComponentProvider() {
+public class MapperComponentProvider extends ClassPathScanningCandidateComponentProvider {
+    public MapperComponentProvider() {
         super(false);
         addIncludeFilter(new AnnotationTypeFilter(Mapper.class, false));
     }
